Add lgbm mode plot of error feature importance and XAI
A notebook that trains a model using a boosting tree model, plots the price of ETH, mean, and Floor price of the Punks, it plots the training of the model and the errors. It shows the top 20 features by feature importance and uses Shap to explain AI estimation and how each feature affects the output.